Изменение : Продолжение 3363, а так же избавились от еще одной неиспользуемой функции;
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@3365 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/transform/MPCVideoDec/DXVADecoderH264.cpp
@@ -111,7 +111,7 @@ void CDXVADecoderH264::CopyBitstream(BYTE* pDXVABuffer, BYTE* pBuffer, UINT& nSi
 				case NALU_TYPE_SLICE:
 				case NALU_TYPE_IDR:
 					// Skip the NALU if the data length is below 0
-					if (Nalu.GetDataLength() < 0) {
+					if ((int)Nalu.GetDataLength() < 0) {
 						break;
 					}
